menambahkan fitur get all dan delete user by admin
