--- routes/auth2.js
@@ -1,118 +1,0 @@
-const express = require('express')
-const router = express.Router()
-const bcrypt = require('bcryptjs')
-const jwt = require('jsonwebtoken')
-
-//import validation
-const { registerValidation } = require('../configs/validation')
-
-// import models
-const Admin = require('../models/Admin')
-const User = require('../models/User')
-const verifyAdmin = require('./verifyAdmin')
-
-// Register
-router.post('/register', async (req, res) => {
-
-    const { error } = registerValidation(req.body)
-    if(error) return res.status(400).json({
-        status: res.statusCode,
-        message: error.details[0].message
-    })
-
-    // if email exist
-    const emailExist = await Admin.findOne({email: req.body.email})
-    if(emailExist) return res.status(400).json({
-        status: res.statusCode,
-        message: 'Email Sudah digunakan !'
-    })
-    
-    // Hash Password
-    const salt = await bcrypt.genSalt(10)
-    const hashPassword = await bcrypt.hash(req.body.password, salt)
-
-    const admin = new Admin({
-        nama: req.body.nama,
-        email: req.body.email,
-        password: hashPassword
-    })
-
-    //create admin
-    try {
-        const saveAdmin = await admin.save()
-        res.json(saveAdmin)
-    }catch(err){
-        res.status(400).json({
-            status: res.statusCode,
-            message: 'Gagal Membuat admin baru'
-        })
-    
-    }
-})
-
-
-// Login 
-router.post('/login', async (req, res) => {
-
-    // if email exist
-    const admin = await Admin.findOne({email: req.body.email})
-    if(!admin) return res.status(400).json({
-        status: res.statusCode,
-        message: 'Email Anda Salah!'
-    })
-
-    // check password
-    const validPwd = await bcrypt.compare(req.body.password, admin.password)
-    if(!validPwd) return res.status(400).json({
-        status: res.statusCode,
-        message: 'Password Anda Salah!'
-    })
-
-    // membuat token menggunkan JWT
-    const token = jwt.sign({ _id: admin._id }, process.env.SECRET_KEY2)
-    res.header('authadmin', token).json({
-        id:admin._id,
-        token: token
-    })
-})
-
-// Update user by admin
-router.put('/updateUser/:id',verifyAdmin, async (req, res) => {
-    try{
-        const userUpdate = await User.updateOne({_id: req.params.id}, {
-            nama: req.body.nama
-        })
-        if(!userUpdate) {
-            res.status(400).json("cek error")
-        } else {
-            const user = await User.findById(req.params.id)
-            res.json(user)
-        }
-    }catch(err){
-        res.status(400).send({message: err})
-    }
-})
-
-// Update data admin
-router.put('/:id',verifyAdmin, async (req, res) => {
-    try{
-        const adminUpdate = await Admin.findByIdAndUpdate({_id: req.params.id}, {
-            nama: req.body.nama
-        });
-
-        if(!adminUpdate) {
-            res.status(400).json("cek error")
-        } else {
-            res.status(200).json({
-                status:res.statusCode,
-                message:{
-                    nama:adminUpdate.nama
-                }
-            });
-        }
-    }catch(err){
-        res.status(400).send({message: err})
-    }
-}),
-
-module.exports = router
